+++ sdk/formrecognizer/azure-ai-formrecognizer/azure/ai/formrecognizer/_models.py
@@ -2905,6 +2905,18 @@ class DocumentStyle:
 
     is_handwritten: Optional[bool]
     """Indicates if the content is handwritten."""
+    similar_font_family: Optional[str]
+    """Visually most similar font from among the set of supported font
+    families, with fallback fonts following CSS convention (ex. 'Arial, sans-serif').
+    """
+    font_style: Optional[str]
+    """Font style. Known values are: "normal", "italic"."""
+    font_weight: Optional[str]
+    """Font weight. Known values are: "normal", "bold"."""
+    color: Optional[str]
+    """Foreground color in #rrggbb hexadecimal format."""
+    background_color: Optional[str]
+    """Background color in #rrggbb hexadecimal format."""
     spans: List[DocumentSpan]
     """Location of the text elements in the concatenated content the style
      applies to."""
@@ -2913,6 +2925,11 @@ class DocumentStyle:
 
     def __init__(self, **kwargs: Any) -> None:
         self.is_handwritten = kwargs.get("is_handwritten", None)
+        self.similar_font_family = kwargs.get("similar_font_family", None)
+        self.font_style = kwargs.get("font_style", None)
+        self.font_weight = kwargs.get("font_weight", None)
+        self.color = kwargs.get("color", None)
+        self.background_color = kwargs.get("background_color", None)
         self.spans = kwargs.get("spans", None)
         self.confidence = kwargs.get("confidence", None)
 
@@ -2920,6 +2937,11 @@ class DocumentStyle:
     def _from_generated(cls, style):
         return cls(
             is_handwritten=style.is_handwritten,
+            similar_font_family=style.similar_font_family,
+            font_style=style.font_style,
+            font_weight=style.font_weight,
+            color=style.color,
+            background_color=style.background_color,
             spans=[DocumentSpan._from_generated(span) for span in style.spans]
             if style.spans
             else [],
@@ -2928,7 +2950,9 @@ class DocumentStyle:
 
     def __repr__(self) -> str:
         return (
-            f"DocumentStyle(is_handwritten={self.is_handwritten}, spans={repr(self.spans)}, "
+            f"DocumentStyle(is_handwritten={self.is_handwritten}, similar_font_family={self.similar_font_family}, "
+            f"font_style={self.font_style}, font_weight={self.font_weight}, color={self.color}, "
+            f"background_color={self.background_color}, spans={repr(self.spans)}, "
             f"confidence={self.confidence})"
         )
 
@@ -2936,6 +2960,11 @@ class DocumentStyle:
         """Returns a dict representation of DocumentStyle."""
         return {
             "is_handwritten": self.is_handwritten,
+            "similar_font_family": self.similar_font_family,
+            "font_style": self.font_style,
+            "font_weight": self.font_weight,
+            "color": self.color,
+            "background_color": self.background_color,
             "spans": [f.to_dict() for f in self.spans]
             if self.spans
             else [],
@@ -2952,6 +2981,11 @@ class DocumentStyle:
         """
         return cls(
             is_handwritten=data.get("is_handwritten", None),
+            similar_font_family=data.get("similar_font_family", None),
+            font_style=data.get("font_style", None),
+            font_weight=data.get("font_weight", None),
+            color=data.get("color", None),
+            background_color=data.get("background_color", None),
             spans=[DocumentSpan.from_dict(v) for v in data.get("spans")]  # type: ignore
             if len(data.get("spans", [])) > 0
             else [],